requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0

# LangChain and Vector Store
langchain==0.0.352
faiss-cpu==1.7.4
chromadb==0.4.22

# Machine Learning
torch==2.1.1
transformers==4.35.2
sentencepiece==0.1.99
accelerate==0.25.0

# Database
sqlalchemy==2.0.23
sqlite3-api==0.1.0
aiosqlite==0.19.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
tqdm==4.66.1
requests==2.31.0
python-multipart==0.0.6

# Development
pytest==7.4.3
black==23.11.0
isort==5.12.0
mypy==1.7.1
